mpc: move MerkleBuoy to merkle module
